fakedb various fixes
git-svn-id: svn://192.168.1.10/public/CoffeeMud@25549 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/fakedb/backend/structure/RecordInfo.java
@@ -20,11 +20,11 @@ limitations under the License.
 */
 public class RecordInfo
 {
-	public int					offset;
+	public long					offset;
 	public int					size;
 	public ComparableValue[]	indexedData	= null;
 
-	public RecordInfo(final int o, final int s)
+	public RecordInfo(final long o, final int s)
 	{
 		offset = o;
 		size = s;
